0.0.0.4
Added a Zombie, Zombie movement towards player, Holding Space slows time, Holding shift speeds up the player and uses a sprint gauge. The sprint gauge and health bar scale with screen width. The sprint gauge recovers by a percentage of the maximum when depleted, but a static amount when accessible.