"Re"writing the code
Not really, I'm just simplifying it as much as I can.
I do what to finish this someday, the sooner the better, but I do think we have to do... better? Idk.

Implemented the basics in terms of compiling into a jar file.

There are two problems to fix... later.
1. Where should additional files be stored? (Such as Keybinds and Settings)
2. To make it an ".exe" for steam, we got two options. Somehow package java into the jar so it can run without needing java installed or compile it into a ".exe" via another program.